All things are migrated from MainActivity.kt to HomeFragment.kt for the working of Viewpager.
UI is neat and clean now what left is logic of viewpager2 and HomeFragment.kt.
