+++ .idea/misc.xml
@@ -5,9 +5,9 @@
       <map>
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/drawable/circle_drawable.xml" value="0.189" />
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout-land/fragment_song.xml" value="0.721875" />
-        <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/activity_main.xml" value="0.22" />
+        <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/activity_main.xml" value="0.33" />
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/fragment_explore_songs.xml" value="0.55" />
-        <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/fragment_home.xml" value="0.25" />
+        <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/fragment_home.xml" value="0.33" />
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/fragment_song.xml" value="0.25" />
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/h_rv_categories.xml" value="0.75" />
         <entry key="..\:/Users/Magician/StudioProjects/Spotify_Clone_App/app/src/main/res/layout/list_item.xml" value="0.55" />
