+++ app/src/main/java/com/plcoding/spotifycloneyt/Model/data/di/AppModule.kt
@@ -1,17 +1,14 @@
 package com.plcoding.spotifycloneyt.Model.data.di
 
-import android.app.Activity
 import android.content.Context
-import android.view.View
 import androidx.navigation.Navigation
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.request.RequestOptions
 import com.plcoding.spotifycloneyt.Model.data.entities.Song
 import com.plcoding.spotifycloneyt.R
-import com.plcoding.spotifycloneyt.View.MainActivity
 import com.plcoding.spotifycloneyt.View.adapters.SongAdapter
-import com.plcoding.spotifycloneyt.View.adapters.SwipeSongsCLicked
+import com.plcoding.spotifycloneyt.View.adapters.SwipeSongAdapter
 import com.plcoding.spotifycloneyt.View.fragments.HomeFragment
 import com.plcoding.spotifycloneyt.other.exoplayer.MusicServiceConnection
 import dagger.Module
@@ -53,12 +50,14 @@ object AppModule {
         }
     }
 
+
     @Singleton
     @Provides
-    fun provideSwipeSongClicked(): SwipeSongsCLicked {
-        return object : SwipeSongsCLicked {
+    fun provideSwipeSongClicked(): SwipeSongAdapter.SwipeSongsCLicked {
+        return object : SwipeSongAdapter.SwipeSongsCLicked {
             override fun SwipeSongCLicked(song: Song) {
-                Navigation.findNavController(HomeFragment().binding.root).navigate(R.id.globalActionToSongFragment)
+                Navigation.findNavController(HomeFragment().binding.root)
+                    .navigate(R.id.globalActionToSongFragment)
             }
         }
     }
