+++ app/src/main/java/com/plcoding/spotifycloneyt/View/MainActivity.kt
@@ -2,49 +2,30 @@ package com.plcoding.spotifycloneyt.View
 
 import android.os.Bundle
 import android.support.v4.media.session.PlaybackStateCompat
-import android.util.Log
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.isVisible
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.Navigation.findNavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.viewpager2.widget.ViewPager2
 import com.bumptech.glide.RequestManager
-import com.google.android.material.snackbar.Snackbar
 import com.plcoding.spotifycloneyt.Model.data.entities.Song
-import com.plcoding.spotifycloneyt.R
-import com.plcoding.spotifycloneyt.View.adapters.SwipeSongAdapter
-import com.plcoding.spotifycloneyt.View.adapters.SwipeSongsCLicked
 import com.plcoding.spotifycloneyt.Viewmodels.MainViewModel
 import com.plcoding.spotifycloneyt.Viewmodels.ViewModelFactory
 import com.plcoding.spotifycloneyt.databinding.ActivityMainBinding
-import com.plcoding.spotifycloneyt.other.Status.*
 import com.plcoding.spotifycloneyt.other.exoplayer.MusicServiceConnection
-import com.plcoding.spotifycloneyt.other.exoplayer.isPlaying
-import com.plcoding.spotifycloneyt.other.exoplayer.toSong
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class MainActivity : AppCompatActivity(), SwipeSongsCLicked {
-
+class MainActivity : AppCompatActivity() {
 
     lateinit var mainViewModel: MainViewModel
-//    lateinit var mainViewModel: MainViewModel
 
-    @Inject
-    lateinit var swipeSongAdapter: SwipeSongAdapter
-
+//    @Inject
+//    lateinit var swipeSongAdapter: SwipeSongAdapter
 
     @Inject
     lateinit var glide: RequestManager
 
     private var curPlayingSong: Song? = null
-
     private var playbackState: PlaybackStateCompat? = null
-
     lateinit var binding: ActivityMainBinding
 
     @Inject
@@ -53,140 +34,17 @@ class MainActivity : AppCompatActivity(), SwipeSongsCLicked {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-
         binding = ActivityMainBinding.inflate(layoutInflater)
         getSupportActionBar()?.hide()
 
         setContentView(binding.root)
-
-
         mainViewModel = ViewModelProvider(
             this,
             ViewModelFactory(musicServiceConnection)
         ).get(MainViewModel::class.java)
-//         mainViewModel = ViewModelProvider(this).get(MainViewModel::class.java)
 
 
-        subscribeToObservers()
-
-        binding.vpSong.adapter = swipeSongAdapter
-        binding.vpSong.registerOnPageChangeCallback(object : ViewPager2.OnPageChangeCallback() {
-            override fun onPageSelected(position: Int) {
-                Toast.makeText(applicationContext, "okay", Toast.LENGTH_SHORT).show()
-                super.onPageSelected(position)
-                if (playbackState?.isPlaying == true) {
-                    mainViewModel.playOrToggleSong(swipeSongAdapter.songs[position])
-                } else {
-                    curPlayingSong = swipeSongAdapter.songs[position]
-                }
-            }
-        })
-        binding.ivPlayPause.setOnClickListener {
-            curPlayingSong?.let {
-                mainViewModel.playOrToggleSong(it, true)
-            }
-        }
-
-//        SwipeSongAdapter.setItemClickListener {
-//        }
-
-
-
-        val navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.navHostFragment) as NavHostFragment
-        val navController = navHostFragment.navController
-
-        setupActionBarWithNavController(navController)
-
-        navController.addOnDestinationChangedListener { _, destination, _ ->
-            when (destination.id) {
-                R.id.songFragment -> hideBottomBar()
-                R.id.homeFragment -> showBottomBar()
-                else -> showBottomBar()
-            }
-        }
     }
 
-    private fun hideBottomBar() {
-        binding.ivCurSongImage.isVisible = false
-        binding.vpSong.isVisible = false
-        binding.ivPlayPause.isVisible = false
-    }
 
-    private fun showBottomBar() {
-        binding.ivCurSongImage.isVisible = true
-        binding.vpSong.isVisible = true
-        binding.ivPlayPause.isVisible = true
-    }
-
-    private fun switchViewPagerToCurrentSong(song: Song) {
-        val newItemIndex = swipeSongAdapter.songs.indexOf(song)
-        if (newItemIndex != -1) {
-            binding.vpSong.currentItem = newItemIndex
-            curPlayingSong = song
-        }
-    }
-
-    private fun subscribeToObservers() {
-        mainViewModel.mediaItems.observe(this) {
-            it?.let { result ->
-                when (result.status) {
-                    SUCCESS -> {
-                        result.data?.let { songs ->
-                            SwipeSongAdapter(glide, songs, this)
-                            if (songs.isNotEmpty()) {
-                                glide.load((curPlayingSong ?: songs[0]).imgUrl)
-                                    .into(binding.ivCurSongImage)
-                            }
-                            switchViewPagerToCurrentSong(curPlayingSong ?: return@observe)
-                        }
-                    }
-                    ERROR -> Unit
-                    LOADING -> Unit
-                }
-            }
-        }
-        mainViewModel.curPlayingSong.observe(this) {
-            if (it == null) return@observe
-
-            curPlayingSong = it.toSong()
-            glide.load(curPlayingSong?.imgUrl).into(binding.ivCurSongImage)
-            switchViewPagerToCurrentSong(curPlayingSong ?: return@observe)
-        }
-        mainViewModel.playbackState.observe(this) {
-            playbackState = it
-            binding.ivPlayPause.setImageResource(
-                if (playbackState?.isPlaying == true) R.drawable.ic_pause else R.drawable.ic_play
-            )
-        }
-        mainViewModel.isConnected.observe(this) {
-            it?.getContentIfNotHandled()?.let { result ->
-                when (result.status) {
-                    ERROR -> Snackbar.make(
-                        binding.rootLayout,
-                        result.message ?: "An unknown error occured",
-                        Snackbar.LENGTH_LONG
-                    ).show()
-                    else -> Unit
-                }
-            }
-        }
-        mainViewModel.networkError.observe(this) {
-            it?.getContentIfNotHandled()?.let { result ->
-                when (result.status) {
-                    ERROR -> Snackbar.make(
-                        binding.rootLayout,
-                        result.message ?: "An unknown error occured",
-                        Snackbar.LENGTH_LONG
-                    ).show()
-                    else -> Unit
-                }
-            }
-        }
-    }
-
-    override fun SwipeSongCLicked(song: Song) {
-        Toast.makeText(this,"SwipeSongAdapter click is working ",Toast.LENGTH_SHORT).show()
-        findNavController(binding.root).navigate(R.id.globalActionToSongFragment)
-    }
 }
