+++ app/src/main/java/com/plcoding/spotifycloneyt/View/fragments/HomeFragment.kt
@@ -1,19 +1,24 @@
 package com.plcoding.spotifycloneyt.View.fragments
 
-import android.annotation.SuppressLint
 import android.os.Bundle
+import android.support.v4.media.session.PlaybackStateCompat
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.viewpager2.widget.ViewPager2
 import com.bumptech.glide.RequestManager
+import com.google.android.material.snackbar.Snackbar
 import com.plcoding.spotifycloneyt.Model.data.entities.Song
 import com.plcoding.spotifycloneyt.R
 import com.plcoding.spotifycloneyt.View.adapters.SongAdapter
+import com.plcoding.spotifycloneyt.View.adapters.SwipeSongAdapter
 import com.plcoding.spotifycloneyt.View.adapters.rv_Genre_Adapter
 import com.plcoding.spotifycloneyt.View.adapters.rv_Language_Adapter
 import com.plcoding.spotifycloneyt.Viewmodels.MainViewModel
@@ -21,15 +26,22 @@ import com.plcoding.spotifycloneyt.Viewmodels.ViewModelFactory
 import com.plcoding.spotifycloneyt.databinding.FragmentHomeBinding
 import com.plcoding.spotifycloneyt.other.Status
 import com.plcoding.spotifycloneyt.other.exoplayer.MusicServiceConnection
+import com.plcoding.spotifycloneyt.other.exoplayer.isPlaying
+import com.plcoding.spotifycloneyt.other.exoplayer.toSong
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked {
+class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked,
+    SwipeSongAdapter.SwipeSongsCLicked {
 
 
     lateinit var mainViewModel: MainViewModel
+    lateinit var binding: FragmentHomeBinding
 
+
+    @Inject
+    lateinit var swipeSongAdapter: SwipeSongAdapter
 
     @Inject
     lateinit var musicServiceConnection: MusicServiceConnection
@@ -40,7 +52,8 @@ class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked 
     @Inject
     lateinit var songAdapter: SongAdapter
 
-    lateinit var binding: FragmentHomeBinding
+    private var curPlayingSong: Song? = null
+    private var playbackState: PlaybackStateCompat? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -50,24 +63,24 @@ class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked 
         binding = FragmentHomeBinding.inflate(inflater, container, false)
         return binding.root
 
+
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
-
-        mainViewModel = ViewModelProvider(requireActivity(),ViewModelFactory(musicServiceConnection)).get(MainViewModel::class.java)
+        mainViewModel =
+            ViewModelProvider(requireActivity(), ViewModelFactory(musicServiceConnection)).get(
+                MainViewModel::class.java
+            )
         setupRecyclerView()
-
-//        subscribeToObservers()
         RV_Language()
         RV_genre()
+        subscribeToObservers()
+        settingUpVP()
+
     }
 
-
-
-     fun setupRecyclerView() {
+    fun setupRecyclerView() {
 
         mainViewModel.mediaItems.observe(viewLifecycleOwner) { result ->
 
@@ -81,11 +94,12 @@ class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked 
                             LinearLayoutManager(
                                 requireContext(),
                                 LinearLayoutManager.VERTICAL,
-                                false)
+                                false
+                            )
                     }
                 }
                 Status.ERROR -> Unit
-                Status.LOADING ->{
+                Status.LOADING -> {
                     binding.allSongsProgressBar.isVisible = true
                 }
             }
@@ -113,7 +127,6 @@ class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked 
     private fun RV_genre() {
         val GenreName = listOf("Love", "Metal", "Romantic", "90's Hits", "OLD is GOLD")
 
-
         val imgList = listOf(
             R.drawable.love,
             R.drawable.metal,
@@ -126,6 +139,129 @@ class HomeFragment : Fragment(R.layout.fragment_home), SongAdapter.SongsCLicked 
         binding.rvLanguage.adapter = adapter
         binding.rvLanguage.layoutManager =
             LinearLayoutManager(requireContext(), LinearLayoutManager.HORIZONTAL, false)
+    }
+
+    //-------------------------------MainActivity data---------------------------------
+
+    private fun subscribeToObservers() {
+        mainViewModel.mediaItems.observe(this) {
+            it?.let { result ->
+                when (result.status) {
+                    Status.SUCCESS -> {
+                        result.data?.let { songs ->
+
+                            // data is going correctly in SwipeSongAdapter
+
+                            Log.d("mytag", "songs count in Mainactivity" + songs.size)
+
+                            SwipeSongAdapter(glide, songs, this)
+
+                            if (songs.isNotEmpty()) {
+                                glide.load((curPlayingSong ?: songs[4]).imgUrl)
+                                    .into(binding.ivCurSongImage)
+                            }
+                            switchViewPagerToCurrentSong(curPlayingSong ?: return@observe)
+                        }
+                    }
+                    Status.ERROR -> Unit
+                    Status.LOADING -> Unit
+                }
+            }
+        }
+        mainViewModel.curPlayingSong.observe(this) {
+            if (it == null) return@observe
+
+            curPlayingSong = it.toSong()
+            glide.load(curPlayingSong?.imgUrl).into(binding.ivCurSongImage)
+            switchViewPagerToCurrentSong(curPlayingSong ?: return@observe)
+        }
+        mainViewModel.playbackState.observe(this) {
+            playbackState = it
+            binding.ivPlayPause.setImageResource(
+                if (playbackState?.isPlaying == true) R.drawable.ic_pause else R.drawable.ic_play
+            )
+        }
+        mainViewModel.isConnected.observe(this) {
+            it?.getContentIfNotHandled()?.let { result ->
+                when (result.status) {
+                    Status.ERROR -> Snackbar.make(
+                        binding.root,
+                        result.message ?: "An unknown error occured",
+                        Snackbar.LENGTH_LONG
+                    ).show()
+                    else -> Unit
+                }
+            }
+        }
+        mainViewModel.networkError.observe(this) {
+            it?.getContentIfNotHandled()?.let { result ->
+                when (result.status) {
+                    Status.ERROR -> Snackbar.make(
+                        binding.root,
+                        result.message ?: "An unknown error occured",
+                        Snackbar.LENGTH_LONG
+                    ).show()
+                    else -> Unit
+                }
+            }
+        }
+    }
+
+    private fun settingUpVP() {
+
+        binding.vpSong.adapter = swipeSongAdapter
+
+        binding.vpSong.registerOnPageChangeCallback(object : ViewPager2.OnPageChangeCallback() {
+            override fun onPageSelected(position: Int) {
+                super.onPageSelected(position)
+                if (playbackState?.isPlaying == true) {
+                    mainViewModel.playOrToggleSong(swipeSongAdapter.songs[position])
+                    //Todo: check if this viewmodel is working or not.
+                } else {
+                    curPlayingSong = swipeSongAdapter.songs[position]
+                }
+            }
+        })
+
+        binding.ivPlayPause.setOnClickListener {
+            curPlayingSong?.let {
+                mainViewModel.playOrToggleSong(it, true)
+            }
+        }
+
+        findNavController().addOnDestinationChangedListener { _, destination, _ ->
+            when (destination.id) {
+                R.id.songFragment -> hideBottomBar()
+                R.id.homeFragment -> showBottomBar()
+                else -> showBottomBar()
+            }
+        }
+    }
+
+    private fun hideBottomBar() {
+        binding.ivCurSongImage.isVisible = false
+        binding.vpSong.isVisible = false
+        binding.ivPlayPause.isVisible = false
+    }
+
+    private fun showBottomBar() {
+        binding.ivCurSongImage.isVisible = true
+        binding.vpSong.isVisible = true
+        binding.ivPlayPause.isVisible = true
+    }
+
+    private fun switchViewPagerToCurrentSong(song: Song) {
+        val newItemIndex = swipeSongAdapter.songs.indexOf(song)
+        if (newItemIndex != -1) {
+            binding.vpSong.currentItem = newItemIndex
+            curPlayingSong = song
+        }
+    }
+
+
+    override fun SwipeSongCLicked(song: Song) {
+        Toast.makeText( requireContext(),"swipeSongAdapter SwipeSong Click listener working \n Heading to SongFrag", Toast.LENGTH_SHORT ).show()
+        findNavController().navigate(R.id.globalActionToSongFragment)
     }
 
     override fun SongCLicked(song: Song) {
